fix: include group/namespace across formats
